The change of traffic accident in Korea and NYC
In 2013, 2014, the traffic accident began to increase. (One of the reasons might be a smartphone.) But the growth rate started to flatten out in 2015, 2016. Public awareness may play an important role, but more effective and lasting environments seem to play a role. Detailed analysis is required before making an exact conclusion. Also, the methodology used in this project can be used to a more important and complicated data analysis in business. 
The plot in the left was captured from an open source, and the plot in the right was made using IDL and the data for the challenge problems.